refactor: remove MeetAi module and migrate participant summary services into the Meeting module.
