--- src/meet-ai/controllers/index.ts
@@ -1,1 +1,0 @@
-export * from './meet-ai.controller';
