--- src/meet-ai/controllers/meet-ai.controller.ts
@@ -1,65 +1,0 @@
-/*
- * @Author: 杨仕明 shiming.y@qq.com
- * @Date: 2026-03-29 20:00:00
- * @LastEditors: 杨仕明 shiming.y@qq.com
- * @LastEditTime: 2026-03-30 14:12:36
- * @FilePath: /nove_api/src/meet-ai/controllers/meet-ai.controller.ts
- * @Description:
- *
- * Copyright (c) 2026 by LuLab-Team, All Rights Reserved.
- */
-
-import {
-  Controller,
-  Get,
-  Post,
-  Body,
-  HttpCode,
-  HttpStatus,
-  Logger,
-  Param,
-} from '@nestjs/common';
-import { ApiTags, ApiBearerAuth, ApiOperation } from '@nestjs/swagger';
-import { NoPermissionRequired } from '@/admin/permission/decorators/permissions.decorator';
-import { ParticipantSummaryService } from '../services';
-import {
-  GenerateParticipantSummaryDto,
-} from '../dto/meet-ai.dto';
-
-@ApiTags('Meet AI')
-@Controller('meet-ai')
-@ApiBearerAuth()
-@NoPermissionRequired()
-export class MeetAiController {
-  private readonly logger = new Logger(MeetAiController.name);
-
-  constructor(
-    private readonly participantSummaryService: ParticipantSummaryService,
-  ) {}
-
-  @Get('health')
-  @HttpCode(HttpStatus.OK)
-  healthCheck() {
-    return {
-      status: 'ok',
-      timestamp: new Date().toISOString(),
-      service: 'meet-ai-service',
-    };
-  }
-
-  @Post('recordings/:recordingId/participant-summaries/generate')
-  @HttpCode(HttpStatus.OK)
-  async generateSummaries(
-    @Param('recordingId') recordingId: string,
-    @Body() dto: GenerateParticipantSummaryDto,
-  ) {
-    return {
-      success: true,
-      message: '参会者总结生成完成',
-      data: await this.participantSummaryService.generateSummaries({
-        recordId: recordingId,
-        ...dto,
-      }),
-    };
-  }
-}
