--- src/meet-ai/dto/index.ts
@@ -1,1 +1,0 @@
-export * from './meet-ai.dto';
