--- src/meet-ai/enums/meet-ai-status.enum.ts
@@ -1,17 +1,0 @@
-/*
- * @Author: 杨仕明 shiming.y@qq.com
- * @Date: 2026-03-29 20:01:33
- * @LastEditors: 杨仕明 shiming.y@qq.com
- * @LastEditTime: 2026-03-29 20:06:44
- * @FilePath: /nove_api/src/meet-ai/enums/meet-ai-status.enum.ts
- * @Description:
- *
- * Copyright (c) 2026 by LuLab-Team, All Rights Reserved.
- */
-
-export enum MeetAiStatus {
-  PENDING = 'pending',
-  PROCESSING = 'processing',
-  COMPLETED = 'completed',
-  FAILED = 'failed',
-}
