--- src/meet-ai/index.ts
@@ -1,1 +1,0 @@
-export * from './meet-ai.module';
