--- src/meet-ai/repositories/index.ts
@@ -1,1 +1,0 @@
-export * from './participant-summary.repository';
