--- src/meet-ai/services/index.ts
@@ -1,1 +1,0 @@
-export * from './participant-summary.service';
