--- src/meet-ai/types/index.ts
@@ -1,1 +1,0 @@
-export * from './meet-ai.types';
