+++ src/meeting/repositories/participant-summary.repository.spec.ts
