+++ src/meeting/services/participant-summary.service.spec.ts
